+++ todo-app/src/models/TaskType.ts
@@ -1,0 +1,5 @@
+export type TaskType = {
+  createdAt: number;
+  isDone: boolean;
+  text: string;
+};
