--- todo-app/src/types/TaskType.tsx
@@ -1,5 +1,0 @@
-export type TaskType = {
-  createdAt: number;
-  isDone: boolean;
-  text: string;
-};
